+++ NextcloudTalk/WebRTC/SignalingMessages/NCUnshareScreenMessage.swift
@@ -1,0 +1,23 @@
+//
+// SPDX-FileCopyrightText: 2020 Nextcloud GmbH and Nextcloud contributors
+// SPDX-License-Identifier: GPL-3.0-or-later
+//
+
+import Foundation
+
+@objcMembers
+public class NCUnshareScreenMessage: NCSignalingMessage {
+
+    public init!(from: String?, to: String?, sid: String?, roomType: String?, payload: [AnyHashable: Any]?) {
+        super.init(from: from, to: to, sid: sid, type: MessageTypeValue.unshareScreen, payload: payload, roomType: roomType, broadcaster: nil)
+    }
+
+    public init(values: [AnyHashable: Any]) {
+        let parsed = NCSignalingMessage.parsedValues(from: values)
+        super.init(from: parsed.from, to: parsed.to, sid: parsed.sid, type: MessageTypeValue.unshareScreen, payload: [:], roomType: parsed.roomType, broadcaster: parsed.broadcaster)
+    }
+
+    public override func messageType() -> NCSignalingMessageType {
+        return .unshareScreen
+    }
+}
